Updating Order and OrderItems endpoints

+++ src/main/java/com/example/ecommerce/Services/CustomerService.java
@@ -5,7 +5,6 @@ import com.example.ecommerce.models.Customer;
 import com.example.ecommerce.repositories.CustomerRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
 import java.util.List;
 import java.util.Optional;
 
@@ -23,7 +22,7 @@ public class CustomerService {
         customerRepository.save(customer);
         return "Customer "+ +customer.getId()+" added successfully";
     }
-    public Customer getCustomerById(Long id) {
+    public Customer getCustomerById(Integer id) {
         customerRepository.findById(id);
         return customerRepository.findById(id).get();
     }
